TASK-3: Пофиксил замечания
